Add new `-i` flag that allows to skip non SemVer tags
Useful for repositories that just have an old tag that is malformed.

Not sure I like the way the shell docs work now but I have not thought
of a decent way to document flags and args at the same time using just
standard lib ways. The usage prints okay, but the `--help` is still
broken only printing the flag information.