Add tests skipping invalid tags
This is a fallout of 6a753e37782ab9309f447dabafc9a70798a76a8d and fixes
the broken tests from that commit. In addition this commit adds a small
number of test cases that check if the behaviour is actually working.

Probably this can be extended if necessary but the logic is not hugely
complicated right now.